buttons in graph sim

+++ include/core_engine/input_handler.h
@@ -13,7 +13,7 @@ class InputHandler
 public:
     InputHandler(const Uint8* keystate, Camera* cam, Display* window, GUI::Canvas* canvas);
     
-    void HandleInput(SDL_Event* e, float delta_time, bool* isGameRunning, float* time_mul = NULL);
+    void HandleInput(SDL_Event* e, float delta_time, bool* isGameRunning, int* time_mul = NULL);
     void HandleCamInput(SDL_Event* e, float delta_time);
 
 protected:
